Extend Firecheckout Compatibility check in its own mixin
