+++ view/frontend/requirejs-config.js
@@ -18,6 +18,9 @@ var config = {
             },
             'Magento_Checkout/js/action/place-order': {
                 'Rvvup_Payments/js/action/place-order-mixin': true
+            },
+            'Rvvup_Payments/js/view/payment/methods/rvvup-paypal': {
+                'Rvvup_Payments/js/view/payment/methods/rvvup-paypal-firecheckout-compatibility-mixin': true
             }
         }
     },
